added PropositionContainer
